Move everything into modules.
This move preserves the names of most classes, so they
can still be referenced from the main puppet repo as before.

In order to avoid naming conflicts, this patch renames
the exim::listserve::private to privateexim::listserve

File references will work just as before -- most files
are not coupled with private manifests so they
remain in the top-level 'files' dir.

Change-Id: I6d7034c941228d80f609eb020a76f20ae6b1d057