Temp. Hack to get tools up and running
secret() causes entire catalog compilation to fail if it can
not find the secret. This breaks tools, since it looks for
this public key on all new instances to setup clush. I've
generated this keypair and thrown away the private key, so
clush won't work until second puppet run.

Will be removed once we've a better solution

Change-Id: I67f485327ebb73d91e84857aabe2b1555a4e22b9