Add Cloud VPS global root key for Alex Monk
I've had cloudinfra admin access for a while (T218448) but have been
occasionally running into projects where I do not have my own access.
Theoretically I could jump onto cloud-cumin and run commands through that but
that feels wrong, not to mention inconvenient.
One of these was T235252#5573628 where a mystery in the tools project
lead to the paws project.
In addition, I've found that sometimes I need to log into the tools project
directly as root without going via my own account (when there are NFS problems)
and so I need to add my key to root for that. I may as well do it globally now
though.

T218448 is relevant here.

This key has come from my Wikimedia Yubikey, and I am PGP signing the commit.
I'm happy to confirm on IRC or upload the public part to a server or Phab or
something as needed.

Change-Id: I5b01125e3ec117e8232ce90c5a9c42233c71e1a2